Arrange the files and code for the new way of fetching API using redux toolkit query
